Create IProfileManager and plumb it through. Start move to folders for ViewModels and Views Seperate Profile to its own file
